unify e2e tests
